+++ src/main/java/lotto/model/DrawResults.java
@@ -1,4 +1,4 @@
-package lotto.model.dto;
+package lotto.model;
 
 import java.util.Collections;
 import java.util.List;
